pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "maestro"
version = "0.1.0"
description = "MAESTRO experiments codebase"
requires-python = ">=3.10"
authors = [{name = "Maestro Team"}]
dependencies = [
    "gymnasium>=0.28",
    "hydra-core>=1.3",
    "matplotlib>=3.7",
    "numpy>=1.24",
    "pandas>=1.5",
    "scikit-learn>=1.2",
    "scipy>=1.9",
    "torch>=2.0",
    "tqdm>=4.65",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
flops = ["ptflops>=0.7"]
dev = ["pytest>=7.4", "black>=23.12.1", "isort>=5.12.0", "flake8>=6.1.0", "mypy>=1.8.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["maestro*"]
